Added randomizing images for emoji, including animated ones
